Add static cast of vector.size() to int

+++ src/SortingAlgorithms.cpp
@@ -13,7 +13,7 @@ namespace SortingAlgorithms {
 		bool swapping = true;
 		while (swapping) {
 			swapping = false;
-			for (int i=0; i<numbers.size() - 1; i++) {
+			for (int i=0; i<static_cast<int> (numbers.size()) - 1; i++) {
 				if (numbers[i] > numbers[i + 1]) {
 					std::swap(numbers[i], numbers[i + 1]);
 					swapping = true;
@@ -25,9 +25,9 @@ namespace SortingAlgorithms {
 	void insertion_sort (std::vector<int>& numbers) {
 		std::vector<int> ordered_numbers = {numbers[0]};
 		bool inserting;
-		for (int i=1; i<numbers.size(); i++) {
+		for (int i=1; i<static_cast<int> (numbers.size()); i++) {
 			inserting = true;
-			for (int j=0; j<ordered_numbers.size(); j++) {
+			for (int j=0; j<static_cast<int> (ordered_numbers.size()); j++) {
 				if (numbers[i] <= ordered_numbers[j]) {
 					ordered_numbers.insert(ordered_numbers.begin() + j, numbers[i]);
 					inserting = false;
